passing through devices

+++ fdf/minilibx-linux/project/scale.c
@@ -28,7 +28,7 @@ void	down_scale(t_data *data)
 		data->vector->scale -= 2;
 }
 
-void	z_up_down(t_data *data, int num)
+void	z_up_down(t_data *data, float num)
 {
 	data->vector->trans_z += num;
 }
